% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BacteriaPI.R
\docType{data}
\name{BacteriaPI}
\alias{BacteriaPI}
\title{pH distribution of 813 bacterial organisms}
\format{A matrix with 813 rows and 768 columns:
\describe{
\item{rowname}{General organism identifier}
\item{colspace}{pH in [0,14]}
}}
\description{
The approximate kernel density estimates of the 813 bacterial 
organisms' isoelectric point (pI) protein distributions.
}
\references{
The authors would like to thank Dr. Chris Knight for providing the original data.
}
