% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFVE.R
\name{GetFVE}
\alias{GetFVE}
\title{Function for converting densities to quantile functions}
\usage{
GetFVE(fpcaObj, dmatrix, dSup, metric = "L2", useAlpha = FALSE, alpha = 0)
}
\arguments{
\item{fpcaObj}{PACE output (FPCA on LQDs)}

\item{dmatrix}{matrix of original densities measures on grid dout, rows correspond to individual densities}

\item{dSup}{support for Density domain - max and min mark the boundary of the support.}

\item{metric}{metric for measuring variance - 'L2' for Euclidean or 'W' for Wasserstein}

\item{useAlpha}{should deregularisation be performed? (default = FALSE)}

\item{alpha}{scalar to deregularise when transforming back to density space with (default = 0)}
}
\value{
FVEvector
}
\description{
Function for converting densities to quantile functions
}
\references{
\cite{Functional Data Analysis for Density Functions by Transformation to a Hilbert space, Alexander Petersen and Hans-Georg Mueller, 2016}
}
\seealso{
\code{\link{lqd2dens},\link{DeregulariseByAlpha}}
}
