% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFVE.R
\name{GetFVE}
\alias{GetFVE}
\title{Function for converting densities to quantile functions}
\usage{
GetFVE(fpcaObj, dmatrix, dSup, alpha = 0.01)
}
\arguments{
\item{fpcaObj}{PACE output (FPCA on LQDs)}

\item{dmatrix}{matrix of densities measures on grid dout, rows correspond to individual densities}

\item{dSup}{support for Density domain - max and min values are assumed to mark the boundary of the support.}

\item{alpha}{scalar to regularise with (default = 0.01)}
}
\value{
FVEvector
}
\description{
Function for converting densities to quantile functions
}
\references{
\cite{Functional Data Analysis for Density Functions by Transformation to a Hilbert space, Alexander Petersen and Hans-Georg Mueller, 2015}
}

