% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDENsample.R
\name{MakeDENsample}
\alias{MakeDENsample}
\title{Convenience function for converting log-quantile functions to densities}
\usage{
MakeDENsample(qmatrix, alpha = 0, dSupUsedInLQ, dSup, useAlpha = FALSE)
}
\arguments{
\item{qmatrix}{Matrix holding the log-quantiles values on [0,1] and assumed to lay strictly on dSupUsedInLQ}

\item{alpha}{Scalar to deregularise the supports with (default=0)}

\item{dSupUsedInLQ}{Support (grid) for Density domain that was used when calculating the LQ projections (maybe different from the original)}

\item{dSup}{Original support (grid) for Density domain of the data}

\item{useAlpha}{Logical indicator to regularise the support so the smallest value of each density is zero}
}
\value{
list with the 'DEN' projected data, and 'dSup' that was originally used for the data.
}
\description{
See 'lqd2dens' and 'RegulariseByAlpha' for more details.
This function will automatically calculate the:sout support points that the 99.5-th percentile (calculated using all available samples) is not positive.
It will then normalise the new densities to integraate to one, regularise by 'alpha' and calculate the LQ projection.
}
\references{
\cite{Functional Data Analysis for Density Functions by Transformation to a Hilbert space, Alexander Petersen and Hans-Georg Mueller, 2015}
}
